requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
anyio==4.6.2.post1
async-timeout==4.0.3
attrs==24.2.0
Automat==24.8.1
beautifulsoup4==4.12.3
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
constantly==23.10.4
cryptography==43.0.3
cssselect==1.2.0
curl_cffi==0.7.3
defusedxml==0.7.1
docker==7.1.0
exceptiongroup==1.2.2
filelock==3.16.1
frozenlist==1.4.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.6
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itemadapter==0.9.0
itemloaders==1.3.2
jmespath==1.0.1
lxml==5.3.0
multidict==6.1.0
packaging==24.1
parsel==1.9.1
propcache==0.2.0
Protego==0.3.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyDispatcher==2.0.7
pyOpenSSL==24.2.1
python-dotenv==1.0.1
PyYAML==6.0.2
queuelib==1.7.0
requests==2.32.3
requests-file==2.1.0
retrying==1.3.4
runstats==2.0.0
scrapinghub==2.4.0
Scrapy==2.11.2
service-identity==24.1.0
shub==2.15.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
tenacity==9.0.0
tldextract==5.1.2
toml==0.10.2
tomli==2.0.2
tqdm==4.55.1
Twisted==24.7.0
typing_extensions==4.12.2
urllib3==2.2.3
w3lib==2.2.1
yarl==1.16.0
zope.interface==7.1.0
zyte-api==0.6.0
